Data Augmentation Script
This script will take files of the power consumption of single appliances and add them together at different times to simulate the aggregate power profile of a house for one day. It will output a CSV file to the same directory where the files are that gives two columns, the first being the total power consumption and the second being a list of appliances that are on at that given time.